Drop import added to TextFields
