+++ VultisigApp/VultisigApp/Views/Components/AddressBook/AddressBookTextField.swift
@@ -51,7 +51,7 @@ struct AddressBookTextField: View {
             }
         }
         .onDrop(of: [.image], isTargeted: $isUploading) { providers -> Bool in
-            handleOnDrop(providers: providers)
+            OnDropQRUtils.handleOnDrop(providers: providers, handleImageQrCode: handleImageQrCode)
         }
     }
     
@@ -200,49 +200,6 @@ struct AddressBookTextField: View {
     private func handleImageQrCode(data: Data) {
         text = String(data: data, encoding: .utf8) ?? ""
         showImagePicker = false
-    }
-    
-    private func handleOnDrop(providers: [NSItemProvider]) -> Bool {
-        guard let provider = providers.first(where: { $0.hasItemConformingToTypeIdentifier("public.image") }) else {
-            print("Invalid file type. Please drop an image.")
-            return false
-        }
-
-        provider.loadDataRepresentation(forTypeIdentifier: "public.image") { data, error in
-            guard let data = data, let image = NSImage(data: data) else {
-                print(error?.localizedDescription ?? "Failed to load image.")
-                return
-            }
-
-            // Extract QR code data from the image
-            if let qrData = extractQRCode(from: image) {
-                handleImageQrCode(data: qrData)
-            } else {
-                print("No QR code detected in the image.")
-            }
-        }
-
-        return true
-    }
-
-    private func extractQRCode(from nsImage: NSImage) -> Data? {
-        guard let tiffData = nsImage.tiffRepresentation,
-              let ciImage = CIImage(data: tiffData) else {
-            return nil
-        }
-
-        let context = CIContext()
-        let detector = CIDetector(ofType: CIDetectorTypeQRCode, context: context, options: [CIDetectorAccuracy: CIDetectorAccuracyHigh])
-
-        if let features = detector?.features(in: ciImage), !features.isEmpty {
-            for feature in features {
-                if let qrFeature = feature as? CIQRCodeFeature, let qrString = qrFeature.messageString {
-                    return Data(qrString.utf8)
-                }
-            }
-        }
-
-        return nil
     }
 }
