+++ VultisigApp/VultisigApp/Views/Components/TextFields/TransactionMemoAddressTextField.swift
@@ -22,6 +22,7 @@ struct TransactionMemoAddressTextField<MemoType: TransactionMemoAddressable>: Vi
     @Binding var isAddressValid: Bool
     @State var showScanner = false
     @State var showImagePicker = false
+    @State var isUploading: Bool = false
     
 #if os(iOS)
     @State var selectedImage: UIImage?
@@ -43,40 +44,36 @@ struct TransactionMemoAddressTextField<MemoType: TransactionMemoAddressable>: Vi
                 }
             }
             
-            ZStack(alignment: .trailing) {
-                if memo.addressFields[addressKey]?.isEmpty ?? true {
-                    placeholder
-                }
-                
-                field
-            }
-            .font(.body12Menlo)
-            .foregroundColor(.neutral0)
-            .frame(height: 48)
-            .frame(maxWidth: .infinity, alignment: .leading)
-            .padding(.horizontal, 12)
-            .background(Color.blue600)
-            .cornerRadius(10)
+            content
+                .font(.body12Menlo)
+                .foregroundColor(.neutral0)
+                .frame(height: 48)
+                .frame(maxWidth: .infinity, alignment: .leading)
+                .background(Color.blue600)
+                .cornerRadius(10)
 #if os(iOS)
-            .sheet(isPresented: $showScanner) {
-                codeScanner
-            }
-            .sheet(isPresented: $showImagePicker, onDismiss: processImage) {
-                ImagePicker(selectedImage: $selectedImage)
-            }
-#elseif os(macOS)
-            .fileImporter(
-                isPresented: $showImagePicker,
-                allowedContentTypes: [UTType.image],
-                allowsMultipleSelection: false
-            ) { result in
-                do {
-                    let qrCode = try Utils.handleQrCodeFromImage(result: result)
-                    handleImageQrCode(data: qrCode)
-                } catch {
-                    print(error)
+                .sheet(isPresented: $showScanner) {
+                    codeScanner
                 }
-            }
+                .sheet(isPresented: $showImagePicker, onDismiss: processImage) {
+                    ImagePicker(selectedImage: $selectedImage)
+                }
+#elseif os(macOS)
+                .fileImporter(
+                    isPresented: $showImagePicker,
+                    allowedContentTypes: [UTType.image],
+                    allowsMultipleSelection: false
+                ) { result in
+                    do {
+                        let qrCode = try Utils.handleQrCodeFromImage(result: result)
+                        handleImageQrCode(data: qrCode)
+                    } catch {
+                        print(error)
+                    }
+                }
+                .onDrop(of: [.image], isTargeted: $isUploading) { providers -> Bool in
+                    OnDropQRUtils.handleOnDrop(providers: providers, handleImageQrCode: handleImageQrCode)
+                }
 #endif
         }
         .onChange(of: memo.addressFields[addressKey]) { oldValue, newValue in
@@ -84,10 +81,32 @@ struct TransactionMemoAddressTextField<MemoType: TransactionMemoAddressable>: Vi
         }
     }
     
-    var placeholder: some View {
-        Text(addressKey.toFormattedTitleCase())
-            .foregroundColor(Color.neutral0)
-            .frame(maxWidth: .infinity, alignment: .leading)
+    var content: some View {
+        field
+            .overlay {
+                ZStack {
+                    if isUploading {
+                        overlay
+                    }
+                }
+            }
+    }
+    
+    var overlay: some View {
+        ZStack {
+            Color.turquoise600.opacity(0.2)
+                .frame(height: 48)
+                .frame(maxWidth: .infinity, alignment: .leading)
+                .cornerRadius(10)
+            
+            RoundedRectangle(cornerRadius: 10)
+                .strokeBorder(Color.turquoise600, style: StrokeStyle(lineWidth: 1, dash: [10]))
+                .padding(5)
+            
+            Text(NSLocalizedString("dropFileHere", comment: ""))
+                .font(.body12MontserratSemiBold)
+                .foregroundColor(.neutral0)
+        }
     }
     
     var field: some View {
@@ -128,6 +147,7 @@ struct TransactionMemoAddressTextField<MemoType: TransactionMemoAddressable>: Vi
 #endif
             addressBookButton
         }
+        .padding(.horizontal, 12)
     }
     
 #if os(iOS)
